requirements.txt
environs
bcrypt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
pydantic
pydantic-settings
alembic
black
pyyaml

flask
marshmallow
